Pull Oxford Stringency Index

API pull of Oxford COVID-19 Government Response Tracker. The stringency index is a composite index across 7 indicators on how governments respond to the COVID epidemic, with scores between 0-100

pull_stringency_index(
  date_start = "2020-01-22",
  date_end = NULL,
  pepfar_only = TRUE
)

Arguments

date_start
start date for stringency index as yyyy-mm-dd, default/min is Jan 22, 2020

date_end
end date for stringency index as yyyy-mm-dd, default is NULL which will run through today

pepfar_only
limit to just PEPFAR countries, default = TRUE

Source

https://covidtracker.bsg.ox.ac.uk/about-api

Value

dataframe of all PEPFAR countries and their stringency index value by day

Details

More info at https://covidtracker.bsg.ox.ac.uk/about-api

Examples

if (FALSE) {
df_stringency <- pull_stringency_index(date_end = "2021-04-01") }